requirements.txt

flask
pillow
pytesseract
werkzeug
